==> _episodes_rmd/03-reshaping-data.Rmd <==
---
title: "Reshaping data with the tidyverse"
author: "Data Carpentry contributors"
teaching: 15
exercises: 15
objectives:
- "Reshape data from wide to long using `pivot_longer()`."
- "Appreciate how data can be transformed from untidy to tidy."

keypoints:
- "Use the `tidyr` package to manipulate dataframes."
- "Pivot functions `pivot_longer()` and `pivot_wider()` can make tidy data."
source: Rmd
---

```{r, echo=FALSE}
source("../bin/chunk-options.R")

```

```{r, echo=FALSE, message=FALSE}
library("tidyverse")

gene_name_file <- "../data/scer-mrna-protein-absolute-estimate.txt"

gene_names <- read_tsv(gene_name_file,comment="#") %>%
    select(Gene=gene,SystematicName=orf)

mRNA_data <- read_tsv("ftp://ftp.ncbi.nlm.nih.gov/geo/series/GSE81nnn/GSE81932/suppl/GSE81932_Dataset01.txt.gz")

names(mRNA_data)[1] <- "SystematicName"

mRNA_data <- left_join(mRNA_data,gene_names)
```

------------

> ### Learning Objectives
>
> * Reshape a tibble
> * Understand pivot_longer and pivot_wider, and how they can help tidy up data


------------


## Reshape/tidy the data

```{r, eval = TRUE, purl = TRUE}
pivot_longer(mRNA_data, 
             cols = !c(SystematicName,Gene), 
             names_to = "Vol", 
             values_to = "log2_ratio")
```

```{r, eval = TRUE, purl = TRUE}
mRNA_data_tidy <- 
    pivot_longer(mRNA_data, 
                 cols = !c(SystematicName,Gene), 
                 names_to = "Vol", 
                 values_to = "log2_ratio")%>%
    separate(Vol,into="Vol_fL") %>%
    mutate(Vol_fL = as.numeric(Vol_fL))

mRNA_data_tidy

```

## Write it out.

```{r, eval = TRUE, purl = TRUE}
write_csv(mRNA_data_tidy, "../data/mRNA_data_tidy_named.csv")

```

```{r, eval = TRUE, purl = TRUE}
mRNA_data_NOP16 <- filter(mRNA_data_tidy, Gene=="NOP16")
mRNA_data_NOP16
```


{% include links.md %}
